Cosponsors file structure
