+++ front/app/containers/InitiativesShow/Cosponsors/index.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import {
+  Box,
+  Title,
+  colors,
+  defaultStyles,
+} from '@citizenlab/cl2-component-library';
+import { useTheme } from 'styled-components';
+
+interface Props {
+  initiativeId: string;
+}
+
+const ListOfCosponsors = ({ initiativeId }: Props) => {
+  const theme = useTheme();
+
+  return (
+    <Box
+      padding="36px"
+      border={`1px solid ${colors.grey300}`}
+      borderRadius={theme.borderRadius}
+      boxShadow={defaultStyles.boxShadow}
+    >
+      <Title variant="h5" as="h2">
+        Cosponsors of this proposal
+      </Title>
+      <ListOfCosponsors initiativeId={initiativeId} />
+    </Box>
+  );
+};
+
+export default ListOfCosponsors;
